// src/app/models/srm/suppliersxpackingviewer.ts
export class SuppliersxPackingViewer{
    idDocumentType: number = -1;
    idSupplierType: number = -1;
    identifier: string = "";
    document: string = "";
    socialReason: string = "";
    supplierType: string = "";
    baseCost: number = -1;
    conversionCost: number = -1;
    lastPurchase: Date= new Date();
}